+++ tests/testthat/test-write_sumstats.R
@@ -4,11 +4,8 @@ test_that("write_sumstats works", {
     is_32bit_windows <- .Platform$OS.type == "windows" ##&&
     ##.Platform$r_arch == "i386"
     if (!is_32bit_windows) {
-        sumstats_dt <- MungeSumstats::formatted_example()
+        sumstats_dt <- MungeSumstats::formatted_example() 
         
-        # res <- MungeSumstats::format_sumstats(path = "~/Downloads/AD_sumstats_Jansenetal_2019sept.txt.gz",
-        #                                ref_genome = "GRCH37",
-        #                                tabix_index=F)
         run_tests <- function(sumstats_dt,
                               fileext,
                               tabix_index=FALSE,